requirements.txt
# Launch1st AI Workflow - Python Dependencies

# Claude API
anthropic>=0.39.0

# Vector Database for RAG
chromadb>=0.4.22

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Utilities
pydantic>=2.5.0
rich>=13.7.0  # Pretty console output
typer>=0.9.0  # CLI framework

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.0

# Optional: Playwright for test execution
playwright>=1.40.0

# Static Analysis (install system-wide or via npm for some)
# pip install semgrep
# npm install -g eslint
# brew install trivy hadolint
